API/requirements.txt
fastapi[all]
uvicorn[standard]
httpx
pandas
scipy
scikit-learn
SQLAlchemy>=1.4
asyncpg
psycopg2-binary
python-dotenv
kafka-python
python-jose
passlib[bcrypt]
